reorganizacion de componentes vue, renombre AdminProductosCat
