+++ package.json
@@ -11,6 +11,7 @@
     },
     "devDependencies": {
         "@popperjs/core": "^2.10.2",
+        "@vue/cli-plugin-router": "~5.0.0",
         "axios": "^0.21",
         "bootstrap": "^5.1.3",
         "case-sensitive-paths-webpack-plugin": "^2.4.0",
@@ -29,6 +30,7 @@
     "dependencies": {
         "@sweetalert2/themes": "^5.0.11",
         "autoprefixer": "10.4.5",
+        "vue-router": "^3.5.4",
         "vue-sweetalert2": "^5.0.5",
         "vuetify": "^2.6.6"
     }
