+++ resources/js/app.js
@@ -8,7 +8,7 @@ require('./bootstrap');
 import Vuetify from '../plugins/vuetify';
 import Vue from 'vue'
 import VueSweetalert2 from 'vue-sweetalert2';
-
+import VueRouter from 'vue-router'
 // If you don't need the styles, do not connect
 import 'sweetalert2/dist/sweetalert2.min.css';
 
@@ -18,6 +18,13 @@ import 'sweetalert2/dist/sweetalert2.min.css';
 
 window.Vue = require('vue').default;
 Vue.use(VueSweetalert2);
+Vue.use(VueRouter);
+
+/**
+ * Aqui crearemos las rutas del frontend para el paner ADMINISTR
+ * Usaremos el plugin router
+*/
+
 
 /**
  * The following block of code may be used to automatically register your
@@ -30,16 +37,13 @@ Vue.use(VueSweetalert2);
 // const files = require.context('./', true, /\.vue$/i)
 // files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(key).default))
 
-Vue.component('example', require('./components/ExampleComponent.vue').default);
+//index de admin
 Vue.component('app-container', require('./components/appContainer.vue').default);
-Vue.component('gestion-productos', require('./components/AdminProductos copy.vue').default);
+Vue.component('gestion-productos', require('./components/AdminProductosCat.vue').default);
 Vue.component('gestion-solicitudes', require('./components/AdminSolicitudes').default);
 
+//Vue.component('admin-productos', require('./components/AdminProductos.vue').default);
 
-//Vue.component('productos', require('./components/Productos.vue').default);
-
-Vue.component('admin-productos', require('./components/AdminProductos.vue').default);
-Vue.component('profile', require('./components/Modal.vue').default);
 
 /**
  * Next, we will create a fresh Vue application instance and attach it to
@@ -50,6 +54,7 @@ Vue.component('profile', require('./components/Modal.vue').default);
 const app = new Vue({
     vuetify: Vuetify,
     vueSweetalert2: VueSweetalert2,
+    vueRouter: VueRouter,
     el: '#app',
 
 });
