--- resources/js/components/ExampleComponent.vue
@@ -1,13 +1,0 @@
-<template>
-    <v-app>
-        <v-btn small color="primary">Prueba</v-btn>
-    </v-app>
-</template>
-
-<script>
-    export default {
-        mounted() {
-            console.log('Component mounted.')
-        }
-    }
-</script>
