+++ resources/js/components/Pruebas/Components.vue
