+++ resources/js/components/Pruebas/Landing.vue
