+++ resources/js/components/Pruebas/Modal.vue
