+++ resources/js/components/Pruebas/Productos.vue
