+++ resources/js/components/Pruebas/Profile.vue
