+++ resources/js/components/Pruebas/Register.vue
