+++ resources/js/components/Pruebas/Starter.vue
